sample_app/requirements.txt
# Browse-to-Test Live Library Verification Requirements
# Install with: pip install -r requirements.txt

# Core library with all features
browse-to-test[all]>=0.2.0

# Optional: Specific AI providers if not included in [all]
# openai>=1.0.0
# anthropic>=0.7.0

# Optional: Specific testing frameworks if not included in [all]  
# playwright>=1.30.0
# selenium>=4.0.0

# Development dependencies for the sample app
pytest>=7.0.0
python-dotenv>=0.19.0 